pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.yapf]
based_on_style = "pep8"
indent_width = 4
column_limit = 100

[tool.isort]
line_length = 100

[project]
name = "c2ot"
version = "1.0.0"
authors = [{ name = "Rex Cheng", email = "hkchengrex@gmail.com" }]
description = ""
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  'torch >= 2.5.1',
  'cython',
  'gitpython >= 3.1',
  'tensorboard >= 2.11',
  'numpy >= 1.21, <2.1',
  'Pillow >= 9.5',
  'opencv-python >= 4.8',
  'scipy >= 1.7',
  'tqdm >= 4.66.1',
  'einops >= 0.6',
  'hydra-core >= 1.3.2',
  'requests',
  'nitrous-ema',
  'hydra_colorlog',
  'tensordict >= 0.6.1',
  'colorlog',
  'open_clip_torch >= 2.29.0',
  'timm >= 1.0.15',
  'python-dotenv',
  'clean-fid',
  'torchdyn >= 1.0.6',
  'pytz',
]

[tool.hatch.build.targets.wheel]
packages = ["c2ot"]
